[Customer] Move CustomerRepository to core
