+++ src/Sylius/Bundle/CoreBundle/Resources/config/app/sylius.yml
@@ -260,6 +260,7 @@ sylius_customer:
         customer:
             classes:
                 model: Sylius\Component\Core\Model\Customer
+                repository: Sylius\Bundle\CoreBundle\Doctrine\ORM\CustomerRepository
                 form:
                     default: Sylius\Bundle\CoreBundle\Form\Type\CustomerType
